+++ rs/moq-ffi/uniffi-bindgen.rs
@@ -1,3 +1,0 @@
-pub fn main() {
-	uniffi::uniffi_bindgen_main();
-}